requirements.txt
# Install in this order — see README.md Setup section for full instructions

# HuggingFace ecosystem
diffusers>=0.32.0
transformers>=4.49.0
accelerate>=1.0.0
huggingface-hub>=0.24.0
datasets>=3.0.0
peft>=0.14.0
trl>=0.15.0

# Quantization
torchao>=0.9.0
bitsandbytes>=0.45.0

# Qwen VL utilities
qwen-vl-utils>=0.0.8
sentencepiece>=0.2.0

# Image processing
Pillow>=10.4.0
numpy>=1.26.0

# Utilities
tqdm>=4.66.0
matplotlib>=3.9.0
ipykernel>=6.0.0
